services/datapipeline_manager/requirements.txt
streamlit==1.37.1
psycopg2-binary==2.9.9
requests==2.32.3
pandas==2.2.2
cryptography==42.0.8
